Factored out InstanceMetadata
